IWriteFile::write now returning size_t. Interface is now like fwrite (from standard c-lib). Also sizeToWrite parameter changed from u32 to size_t.
git-svn-id: svn://svn.code.sf.net/p/irrlicht/code/trunk@5334 dfc29bdd-3216-0410-991c-e03cc46cb475

+++ source/Irrlicht/CImageWriterBMP.cpp
@@ -105,7 +105,7 @@ bool CImageWriterBMP::writeImage(io::IWriteFile* file, IImage* image, u32 param)
 	u32 row_stride = (pixel_size * imageHeader.Width);
 
 	// length of one row in bytes, rounded up to nearest 4-byte boundary
-	s32 row_size = ((3 * imageHeader.Width) + 3) & ~3;
+	size_t row_size = ((3 * imageHeader.Width) + 3) & ~3;
 
 	// allocate and clear memory for our scan line
 	u8* row_pointer = new u8[row_size];
